pyproject.toml
[project]
name = "bookkeeper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.116.2",
    "httpx>=0.28.1",
    "langfuse>=3.5.0",
    "litellm>=1.77.3",
    "pydantic>=2.11.9",
    "retrying>=1.4.2",
    "strands-agents>=1.9.0",
    "strands-agents-tools>=0.2.8",
    "uvicorn>=0.35.0",
]
